Layout updates and add permission error view.
